Refactor meals and likes js files into respective folders
