+++ src/modules/meals/loadMeals.js
@@ -1,6 +1,6 @@
 import getMeals from './getMeals.js';
 import showMeal from './showMeal.js';
-import loadLikes from './loadLikes.js';
+import loadLikes from '../likes/loadLikes.js';
 
 const loadMeals = async () => {
   const meals = await getMeals();
